chore: move prompt, findProjectRoot, toBigNumberString
